Create an SBOM attestation
A Software Bill of Materials (SBOM) attestations describe what software
artifacts an image contains, and artifacts used to create the image.

Metadata included in an SBOM for describing software artifacts may
include:

- Name of the artifact
- Version
- License type
- Authors
- Unique package identifier

There are benefits to indexing contents of an image during the build, as
opposed to scanning a final image. When scanning happens as part of the
build, you're able to detect software you use to build the image, that
may not show up in the final image.

The SBOMs generated by BuildKit follow the SPDX standard. SBOMs attach
to the final image as a JSON-encoded SPDX document.